MFC r361346: Retain only mutually supported TCP options after simultaneous SYN
When receiving a parallel SYN in SYN-SENT state, remove all the
options only we supported locally before sending the SYN,ACK.

This addresses a consistency issue on parallel opens.

Also, on such a parallel open, the stack could be coaxed into
running with timestamps enabled, even if administratively disabled.

This does NOT contain the merge of the change to BBR since at this
time that code does not exist in stable/12, and there is no plan to
merge BBR(v1) to stable/12.

Reviewed by:	tuexen (mentor)
Approved by:	tuexen (mentor), rgrimes (mentor, blanket)
Sponsored by:	NetApp, Inc.
Differential Revision:	https://reviews.freebsd.org/D23371


git-svn-id: https://svn.freebsd.org/base/stable/12@361841 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f