MFC r361758
netmap: vtnet: add vtnrx_nm_refill index to receive queues

The new index tracks the next netmap slot that is going
to be enqueued into the virtqueue. The index is necessary
to prevent the receive VQ and the netmap rx ring from going
out of sync, considering that we never enqueue N slots, but
at most N-1. This change fixes a bug that causes the VQ
and the netmap ring to go out of sync after N-1 packets
have been received.


git-svn-id: https://svn.freebsd.org/base/stable/12@362024 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f