Use the correct address when creating pci resources
When the PCI and CPU physical addresses are identical it doesn't matter
which is used to create the resources, however on some systems, e.g.
qemu armv7 virt, they are different. This leads to a panic as we try to
map the wrong physical address into the kernel address space.

Reported by:	Jenkins via trasz
Sponsored by:	Innovate UK


git-svn-id: https://svn.freebsd.org/base/head@362397 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f